+++ example/filelist.go.sample
