+++ example/fs.go
@@ -2,7 +2,7 @@ package main
 
 import (
 	"fmt"
-	"../fileapi"
+	"fileapi"
 )
 
 func main() {
